refactor: TargetLocator
