+++ FairyTaleDefender/Assets/_Game/ScriptableObjects/Weapons/Targeting/ClosestTargetType.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60b758fae1fe95b44b420de43903e9d1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
