+++ FairyTaleDefender/Assets/_Game/ScriptableObjects/Weapons/Targeting/RandomTargetType.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb9469c0d14e3e84cb027962df1638e0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
