+++ FairyTaleDefender/Assets/_Game/Scripts/Runtime/Entities/Weapons/Targeting/ScriptableObjects/ClosestTargetTypeSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 36139b81781c48e79b91ba66d83fc840
+timeCreated: 1681041094
