+++ FairyTaleDefender/Assets/_Game/Scripts/Runtime/Entities/Weapons/Targeting/ScriptableObjects/RandomTargetTypeSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 975252c5e4014f779bffc0a2d3a5306b
+timeCreated: 1681040971
