+++ FairyTaleDefender/Assets/_Game/Scripts/Runtime/Entities/Weapons/Targeting/ScriptableObjects/TargetTypeSO.cs
@@ -1,0 +1,10 @@
+using BoundfoxStudios.FairyTaleDefender.Infrastructure;
+using UnityEngine;
+
+namespace BoundfoxStudios.FairyTaleDefender.Entities.Weapons.Targeting.ScriptableObjects
+{
+	public abstract class TargetTypeSO : ScriptableObject
+	{
+		public abstract Collider GetTargetNonAlloc(Vector3 weaponPosition, NoAllocArrayResult<Collider> targets);
+	}
+}
