--- FairyTaleDefender/Assets/_Game/Scripts/Runtime/Entities/Weapons/Targeting/TargetType.cs
@@ -1,8 +1,0 @@
-namespace BoundfoxStudios.FairyTaleDefender.Entities.Weapons.Targeting
-{
-	public enum TargetType
-	{
-		Closest = 0,
-		Random = 1,
-	}
-}
