--- FairyTaleDefender/Assets/_Game/Scripts/Runtime/Entities/Weapons/Targeting/TargetType.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 96c9c3b684fd4907826b3c899ec15ae8
-timeCreated: 1674053961
